requirements.txt
psycopg-binary
numpy
tensorflow
fire
tqdm
orjson
requests